[WFLY-266] Notification Support
* expose the NotificationSupport directly from the OperationContext. the
  instance is held by the OperationContext's modelController
