+++ controller/src/test/java/org/jboss/as/controller/notification/NotificationCompositeOperationTestCase.java
@@ -90,7 +90,7 @@ public class NotificationCompositeOperationTestCase extends AbstractControllerTe
                             @Override
                             public void handleResult(OperationContext.ResultAction resultAction, OperationContext context, ModelNode operation) {
                                 Notification notification = new Notification(MY_TYPE, pathAddress(operation.get(OP_ADDR)), operation.get("param").asString());
-                                NotificationService.emitNotification(context, notification);
+                                context.getNotificationSupport().emit(notification);
                             }
                         });
                     }
@@ -102,10 +102,8 @@ public class NotificationCompositeOperationTestCase extends AbstractControllerTe
                 new OperationStepHandler() {
                     @Override
                     public void execute(OperationContext context, ModelNode operation) throws OperationFailedException {
-                        ServiceController<?> notificationService = context.getServiceRegistry(false).getService(NotificationService.SERVICE_NAME);
-                        NotificationSupport notificationSupport = NotificationSupport.class.cast(notificationService.getValue());
                         PathAddress source = pathAddress(operation.get(OP_ADDR));
-                        notificationSupport.registerNotificationHandler(source, delegatingNotificationHandler, delegatingNotificationHandler);
+                        context.getNotificationSupport().registerNotificationHandler(source, delegatingNotificationHandler, delegatingNotificationHandler);
                         context.stepCompleted();
                         notificationHandlerRegisteredLatch.get().countDown();
                     }
@@ -117,17 +115,13 @@ public class NotificationCompositeOperationTestCase extends AbstractControllerTe
                 new OperationStepHandler() {
                     @Override
                     public void execute(OperationContext context, ModelNode operation) throws OperationFailedException {
-                        ServiceController<?> notificationService = context.getServiceRegistry(false).getService(NotificationService.SERVICE_NAME);
-                        NotificationSupport notificationSupport = NotificationSupport.class.cast(notificationService.getValue());
                         PathAddress source = pathAddress(operation.get(OP_ADDR));
-                        notificationSupport.unregisterNotificationHandler(source, delegatingNotificationHandler, delegatingNotificationHandler);
+                        context.getNotificationSupport().unregisterNotificationHandler(source, delegatingNotificationHandler, delegatingNotificationHandler);
                         context.stepCompleted();
                         notificationHandlerUnregisteredLatch.get().countDown();
                     }
                 }
         );
-
-        NotificationService.installNotificationService(getContainer().subTarget());
     }
 
     @Ignore
